tools/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
elastic-transport==8.15.1
elasticsearch==8.16.0
et_xmlfile==2.0.0
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.1.3
openai==0.28.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.241126
pdf2image==1.17.0
pillow==11.0.0
propcache==0.2.0
PyJWT==2.10.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
tqdm==4.67.1
types-pytz==2024.2.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.0
